More clean code
