+++ src/components/MobileMetricCard/MobileMetricCard.js
@@ -3,7 +3,7 @@ import { graphql } from 'react-apollo'
 import cx from 'classnames'
 import Label from '@santiment-network/ui/Label'
 import { formatTooltipValue } from '../../ducks/SANCharts/CustomTooltip'
-import { Metrics } from '../../ducks/SANCharts/utils'
+import { Metrics } from '../../ducks/SANCharts/data'
 import PercentChanges from '../PercentChanges'
 import { METRIC_ANOMALIE_QUERY } from '../../ducks/GetTimeSeries/queries/metric_anomaly_query'
 import styles from './MobileMetricCard.module.scss'
