+++ src/ducks/SANCharts/ChartMetricSelector.js
@@ -9,7 +9,8 @@ import Button from '@santiment-network/ui/Button'
 import MetricExplanation from './MetricExplanation'
 import ExplanationTooltip from '../../components/ExplanationTooltip/ExplanationTooltip'
 import { PROJECT_METRICS_BY_SLUG_QUERY } from './gql'
-import { Metrics, Events, getMarketSegment } from './utils'
+import { getMarketSegment } from './utils'
+import { Metrics, Events } from './data'
 import styles from './ChartMetricSelector.module.scss'
 
 const NO_GROUP = '_'
