+++ src/ducks/SANCharts/ChartPage.js
@@ -8,7 +8,8 @@ import GetTimeSeries from '../../ducks/GetTimeSeries/GetTimeSeries'
 import { ERRORS } from '../GetTimeSeries/reducers'
 import Charts from './Charts'
 import Header from './Header'
-import { Metrics, Events, getMarketSegment } from './utils'
+import { getMarketSegment } from './utils'
+import { Metrics, Events } from './data'
 import { getNewInterval, INTERVAL_ALIAS } from './IntervalSelector'
 import UpgradePaywall from './../../components/UpgradePaywall/UpgradePaywall'
 import { getIntervalByTimeRange } from '../../utils/dates'
