+++ src/ducks/Signals/chart/SignalPreview.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { ReferenceLine } from 'recharts'
 import Loader from '@santiment-network/ui/Loader/Loader'
 import { getMetricsByType, getTimeRangeForChart } from '../utils/utils'
-import { Metrics } from '../../SANCharts/utils'
+import { Metrics } from '../../SANCharts/data'
 import GetTimeSeries from '../../GetTimeSeries/GetTimeSeries'
 import ChartWidget from '../../SANCharts/ChartPage'
 import VisualBacktestChart from './VisualBacktestChart'
