+++ src/pages/Chart/index.js
@@ -3,7 +3,7 @@ import GA from 'react-ga'
 import { parse } from 'query-string'
 import { graphql, Query } from 'react-apollo'
 import ChartWidget from '../../ducks/SANCharts/ChartPage'
-import { Metrics } from '../../ducks/SANCharts/utils'
+import { Metrics } from '../../ducks/SANCharts/data'
 import InsightsWrap from '../../components/Insight/InsightsWrap'
 import AnonBannerCardB from '../../components/Banner/AnonBanner/AnonBannerCardB'
 import { ALL_INSIGHTS_BY_PAGE_QUERY } from '../../queries/InsightsGQL'
